requirements.txt
accelerate>=0.24.0
transformers>=4.36.0
datasets>=2.15.0
peft>=0.7.0
python-dotenv>=1.0.0
torch>=2.0.0
pyyaml>=6.0
jinja2>=3.1.0
rich>=13.6.0
matplotlib>=3.7.0
seaborn>=0.12.0
